+++ test/test_effectscatter.py
@@ -74,5 +74,5 @@ def test_effectscatter_multiple_symbol_type():
 
 def test_effectscatter_splitline():
     es = EffectScatter("动态散点图各种图形示例")
-    es.add("", [10], [10],)
+    es.add("", [10], [10])
     assert es.options["xAxis"][0]["splitLine"]["show"] is True