+++ pyecharts/charts/effectscatter.py
@@ -15,6 +15,7 @@ class EffectScatter(Scatter):
 
     def add(self, *args, **kwargs):
         self.__add(*args, **kwargs)
+        return self
 
     def __add(self, name, x_axis, y_axis, symbol_size=10, **kwargs):
         """
@@ -34,6 +35,8 @@ class EffectScatter(Scatter):
         chart = self._get_all_options(**kwargs)
 
         xaxis, yaxis = chart["xy_axis"]
+        # show split line, because by default split line is hidden for xaxis
+        xaxis[0]["splitLine"]["show"] = True
         self._option.update(xAxis=xaxis, yAxis=yaxis)
         self._option.get("legend")[0].get("data").append(name)